requirements.txt
flask==3.0.0
google-generativeai==0.5.4
python-dotenv==1.0.1
gunicorn==21.2.0